Fixes for parameter lists
Now handling of multiple variables is not done in the grammar or the
parse tree. It must be done by the semantic analyzer. An appropriate
warning is also posted in the class for inbuilt types.
Renamed classes: GoType -> GoBaseType, GoInbuiltType -> GoType, to
account for the above change.
Makefile updated to clean __pycache__ and *.pyc
Added testcases for the range and switch-case